Authenticate withdrawal requests
Use REST controller name + action

[Finishes #160533509]
